chore: chatbox renamed as GameRoom
