+++ client/components/index.js
@@ -10,4 +10,4 @@ export {Start} from './StartForm'
 export {default as Whiteboard} from './Whiteboard'
 export {default as Home} from './Home'
 export {default as Instructions} from './Instructions'
-export {default as ChatBox} from './ChatBox'
+export {default as GameRoom} from './GameRoom'
